LPS-140039 Move dragTypes to data-engine-js-components-web
